build: add static_cast to fix Clang debug build comparison warning
Fix signed/unsigned comparison warning in BGSDefaultObjectManager that
only appears in Clang debug builds:

  error: comparison of integers of different signs: 'std::size_t'
  (aka 'unsigned long long') and 'int' [-Werror,-Wsign-compare]

Add static_cast<std::size_t> to the Relocate() call to match idx type.

+++ src/RE/B/BGSDefaultObjectManager.cpp
@@ -29,7 +29,7 @@ namespace RE
 		if (idx == kInvalid) {
 			return nullptr;
 		}
-		assert(idx < Relocate(364, 364, 369));
+		assert(idx < static_cast<std::size_t>(Relocate(364, 364, 369)));
 		return (&RelocateMember<bool>(this, 0xB80, 0xBA8))[idx] ?
 		           &(&RelocateMember<TESForm*>(this, 0x20, 0x20))[idx] :
 		           nullptr;
